requirements.txt
streamlit==1.56.0
anthropic>=0.28.0
PyGithub>=2.3.0
python-dotenv>=1.0.0
Authlib>=1.3.2
requests>=2.31.0
boto3>=1.34.0
google-auth>=2.28.0
google-cloud-functions>=1.16.0
google-cloud-iam>=2.15.0
google-cloud-pubsub>=2.21.0
google-cloud-run>=0.10.0
